+++ driver-core/src/main/java/com/datastax/driver/core/Connection.java
@@ -588,8 +588,7 @@ class Connection extends com.datastax.cassandra.transport.Connection
         }
 
         // Make sure we don't print huge responses in debug/error logs.
-        private String asDebugString(Object obj)
-        {
+        private String asDebugString(Object obj) {
             if (obj == null)
                 return "null";