+++ driver-core/src/main/java/com/datastax/driver/core/DefaultResultSetFuture.java
@@ -221,4 +221,3 @@ class DefaultResultSetFuture extends SimpleFuture<ResultSet> implements ResultSe
         }
     }
 }
-