+++ driver-core/src/main/java/com/datastax/driver/core/policies/RoundRobinPolicy.java
@@ -23,8 +23,11 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import com.google.common.collect.AbstractIterator;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.datastax.driver.core.Cluster;
+import com.datastax.driver.core.ConsistencyLevel;
 import com.datastax.driver.core.Host;
 import com.datastax.driver.core.HostDistance;
 import com.datastax.driver.core.Query;
@@ -43,8 +46,12 @@ import com.datastax.driver.core.Query;
  */
 public class RoundRobinPolicy implements LoadBalancingPolicy {
 
+    private static final Logger logger = LoggerFactory.getLogger(RoundRobinPolicy.class);
+
     private final CopyOnWriteArrayList<Host> liveHosts = new CopyOnWriteArrayList<Host>();
     private final AtomicInteger index = new AtomicInteger();
+
+    private volatile boolean hasLoggedLocalCLUse;
 
     /**
      * Creates a load balancing policy that picks host to query in a round robin
@@ -86,6 +93,13 @@ public class RoundRobinPolicy implements LoadBalancingPolicy {
      */
     @Override
     public Iterator<Host> newQueryPlan(Query query) {
+
+        if (!hasLoggedLocalCLUse && query.getConsistencyLevel() != null && query.getConsistencyLevel().isDCLocal()) {
+            hasLoggedLocalCLUse = true;
+            logger.warn("Detected request at Consistency Level {} but the non-DC aware RoundRobinPolicy is in use. "
+                      + "It is strongly advised to use DCAwareRoundRobinPolicy if you have multiple DCs/use DC-aware consistency levels "
+                      + "(note: this message will only be logged once)", query.getConsistencyLevel());
+        }
 
         // We clone liveHosts because we want a version of the list that
         // cannot change concurrently of the query plan iterator (this