+++ driver-core/CHANGELOG.rst
@@ -1,6 +1,17 @@
 CHANGELOG
 =========
 
+1.0.7:
+------
+
+- [new] Add method to check if a Cluster/Session instance has been closed already (JAVA-114)
+- [new] Expose the name of the partitioner in use in the cluster metadata (JAVA-179)
+- [new] Add new WhiteListPolicy to limit the nodes connected to a particular list
+- [improvement] Do not hop DC for LOCAL_* CL in DCAwareRoundRobinPolicy (JAVA-289)
+- [bug] Revert back to longs for dates in the query builder (JAVA-313)
+- [bug] Don't reconnect to nodes ignored by the load balancing policy (JAVA-314)
+
+
 1.0.6:
 ------